% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mins_to_hours.R
\name{mins_to_hours}
\alias{mins_to_hours}
\title{mins_to_hours}
\usage{
mins_to_hours(mins)
}
\arguments{
\item{mins}{is a numeric, atomic, or vector. I named it "mins" because it should be an input of a number of minutes which will be converted to a number of hours.}
}
\value{
a numeric of the value divided by 60
}
\description{
mins_to_hours returns the numeric value converted from a measurement in minutes to a measurement in hours
}
\examples{
mins_to_hours(60)
mins_to_hours(1:5)
mins_to_hours(0)
}
